requirements.txt
# ==================== 核心框架 ====================
# Qwen-Agent 核心框架（包含 RAG、代码解释器、GUI 功能）
qwen-agent[rag,code_interpreter,gui]>=0.0.1
json5>=0.9.0
neo4j==5.22.0
faiss-cpu>=1.7.4
numpy>=1.21.0
dashscope>=1.10.0
requests>=2.28.0


